requirements.txt
neo4j==4.1.0
graphql-core==3.0.5
pydash==4.8.0
uvicorn~=0.16.0
setuptools~=59.1.1
ariadne~=0.14.0